// src/app/model/points.model.ts
export class Points {
    id:                     number;
    points:                 string;
    nombre_carpeta_destino: string;
    ip:                     string;
    ciudad:                 string;
    dateCreated:            string;
    user_correo: string;
}

// Converts JSON strings to/from your types
// and asserts the results of JSON.parse at runtime